+++ src/App.jsx
@@ -4,7 +4,7 @@ import { Routes, Route } from "react-router-dom";
 // Components
 import Home from "./pages/Home";
 // import Play from "./pages/Play";
-import Play from "./pages/PlayWithDateTime";
+import Play from "./pages/Play";
 import Setting from "./pages/Setting";
 
 // Context
