+++ src/assets/tracks.js
@@ -12,6 +12,7 @@ import ten from "./ten.wav";
 import mixdown from "./mixdown.wav";
 
 export const tracks = [
+  // { title: "mixdown", src: mixdown},
   { title: "one", src: one },
   { title: "two", src: two },
   { title: "three", src: three },
